requirements.txt
openai
sqlalchemy
streamlit
faiss-cpu
numpy
pandas
